pyproject.toml
[project]
name = "grid-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "farsounder>=0.1.1",
    "rerun-sdk>=0.30.1",
    "utm>=0.8.1",
]
